Add RSS-feed for blog posts
We use the plugin `gatsby-plugin-feed` to automatically generate an
RSS-feed from the Markdown-files:

https://www.gatsbyjs.com/plugins/gatsby-plugin-feed/

We need to override the default serialize-function of this plugin to add
the `/blog` to the URL of the blog posts. The query was taken from the
documentation as-is. The resulting feed is saved as `/blog.xml`.

Note that the feed is only regenerated during a build, not during
development. To inspect any changes, run `npm run build` before running
`npm start`.